guestboard/feat: guestboard CRUD 기능 구현

+++ src/main/java/com/example/demo/guestboard/dto/InfiniteScroll.java
@@ -4,6 +4,6 @@ import lombok.Data;
 
 @Data
 public class InfiniteScroll {
-  private int offset;
-  private int limit;
+  private Long offset;
+  private Long limit;
 }
